+++ src/main/java/tarehart/rlbot/math/DistanceTimeSpeed.kt
@@ -2,4 +2,4 @@ package tarehart.rlbot.math
 
 import tarehart.rlbot.time.Duration
 
-class DistanceTimeSpeed(val distance: Double, val time: Duration, val speed: Double)
+class DistanceTimeSpeed(val distance: Float, val time: Duration, val speed: Float)
