+++ src/main/java/tarehart/rlbot/math/Rectangle.kt
@@ -6,6 +6,6 @@ import java.awt.geom.Rectangle2D
 class Rectangle(val awtRectangle: Rectangle2D) {
 
     operator fun contains(test: Vector2): Boolean {
-        return awtRectangle.contains(test.x, test.y)
+        return awtRectangle.contains(test.x.toDouble(), test.y.toDouble())
     }
 }
