+++ src/main/java/tarehart/rlbot/math/vector/Spin.kt
@@ -2,7 +2,7 @@ package tarehart.rlbot.math.vector
 
 
 
-class Spin(val pitchRate: Double, val yawRate: Double, val rollRate:Double) {
+class Spin(val pitchRate: Float, val yawRate: Float, val rollRate:Float) {
 
     constructor(angularVelLocal: Vector3): this(-angularVelLocal.y, -angularVelLocal.z, angularVelLocal.x)
 }
