+++ src/main/java/tarehart/rlbot/routing/DistanceDuration.kt
@@ -3,4 +3,4 @@ package tarehart.rlbot.routing
 import tarehart.rlbot.time.Duration
 
 
-class DistanceDuration(val distance:Double, val duration: Duration)
+class DistanceDuration(val distance: Float, val duration: Duration)
